Description: fix the problems about wearable devices compile alarms IssueNo: https://gitee.com/openharmony/graphic_ui/issues/I53JQ8 Feature or Bugfix: Bugfix Binary Source:No Signed-off-by: liuyuxiang <liuyuxiang7@huawei.com>

+++ frameworks/src/core/wrapper/js.h
@@ -84,7 +84,7 @@ public:
 
 class JSNumber final : public MemoryHeap {
 public:
-    static JSValue Create(uint32_t number);
+    static JSValue Create(double number);
     static bool Is(JSValue target);
     static double Value(JSValue value);
 };
